+++ Sources/LiveLocalizationCore/LiveLocalizationVersion.swift
@@ -1,5 +1,5 @@
 import Foundation
 
-public enum DebugLocalizationVersion {
+public enum LiveLocalizationVersion {
     public static let current = "0.3.0-dev"
 }